+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+New Features
+~~~~~~~~~~~~
+
+* finagle-base-http: Add support for new "b3" tracing header. ``PHAB_ID=D334419``
+
 Runtime Behavior Changes
 ~~~~~~~~~~~~~~~~~~~~~~~~